Add a file input to attachments' resolution view
